+++ changelog.md
@@ -1,16 +1,18 @@
 # Changelog
 All changes that affects usability to this project will be documented here.
 ## 1.1.2 - *WIP*
- - Under development
+ - MDT tool under development
+ - Added two new cards at main menu
 ## 1.1.1 - *09, october 2022*
 `Added`
 - Support for AEV to OBJ conversion (export and reimport), best viewed on blender
-- Added new menu buttons for ETS and ITA
+- Added new menu buttons for all 3 tools 
 - Added better visual feedback when saving file
 
 `Fixed`
 - Fix first event showing wrong room 
 - Fix header position on ETS and ITA tools
+- Fix lower and higher limits not updating
 ## 1.1.0 - *07, october 2022*
 `Added`
 - New redesigned AEV tool